Initial integration of SDP Control Client into fwknop server
Modify the server as well as the control client to work together. Server
can now use the fwknop client to send a SPA packet, connect to the
controller, receive a credential update, and receive its access data
from the controller. This commit does not yet enable the server to
continue listening to the controller for further updates.