Modify test suite to disable SDP control client by default
This commit does not add tests to exercise SDP control client
functionality. It updates the suite to disable the control client by
default so that the pre-existing tests still work. Also, working through
the tests, it became clear that we don't need to protect the access data
hash table with a mutex when reading the access data from a file because
it means there is no threading and no chance of a race condition.