Modify controller interaction workflow
When connecting to controller, controller will check if a credential
update is required based on its own settings and either send new
credentials or a "credentials_good" message. Client always waits for one
of those two messages. If new credentials were NOT sent and the client
feels they are due, it will request them. After getting those, if it's a
gateway and the time is right, it will request access data to populate
the access data hash table.