Enable config dump with SDP CTRL Client
Added fwknopd.conf setting to direct config-dump output to a specified
file. In conjunction with this, modified server so that, when configured
to run SDP CTRL Client, 'fwknopd --dump-config' now sends a signal to
the running daemon to execute the dump, which writes to the specified
file.

SECURITY WARNING: In SDP mode, dump-config is now printing SPA keys for
the time being to help debug other issues.