Add service list to gate
Add capability to gate to receive list of services from controller along
with info to connect to them. Modify access stanzas to list service IDs
that a client can access. All of this is incremental, meaning the data
structures are added and the ability to get this info from the
controller, but SPA packets are not yet modified to request a service id
and the server doesn't yet know how to handle such SPA packets.