Open tunnel service when SPA arrives
When a valid SPA packet is received, if tunnel service is enabled,
instead of opening a port to the requested service, a notice is
submitted for the tunnel manager to expect an incoming connection and
then the port for the tunnel service is opened. For now, in the case of
SPA to reach the controller, the old behavior is followed (open the
controller port, not the tunnel service port). Tests were updated to
disable the tunnel service so that the expected port is opened for each
of those tests.